# src/main/java/com/training/studies/entities/Payment.java
package com.training.studies.entities;

import jakarta.persistence.*;
import lombok.*;

import java.time.LocalDate;

@Entity
@AllArgsConstructor @NoArgsConstructor @Getter @Setter @ToString @Builder
public class Payment {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "id", nullable = false)
    private Long id;


    private LocalDate date;

    private double amount;

    private PaymentType type;

    private PaymentStatus status ;

    private String file ;

    @ManyToOne
    private Student student ;

}